requirements.txt
requests>=2.31.0
feedparser>=6.0.10
datasketch>=1.6.4
pandas>=2.1.0
streamlit>=1.32.0
pyyaml>=6.0.1
plotly>=5.20.0
